Fix errors/warnings

+++ C++/RMutex.cpp
@@ -4,7 +4,7 @@ RMutexArray::RMutexArray(size_t sz, const std::string& initial_name)
   : size (sz),
 	 mutexes (new RMutex[sz])
 {
-  for (int i = 0; i < sz; i++)
+  for (size_t i = 0; i < sz; i++)
 	 mutexes[i].set_name(initial_name);
 }
